requirements.txt
auditwheel
av
chartio
Cython
defusedxml
docutils
h5py
imantics
ipython
Jinja2
lmdb
matplotlib
numpy
opencv_python
pandas
Pillow
pytest
pytorch_sphinx_theme
PyYAML
requests
scipy
setuptools
Sphinx
torch
torchdata
tqdm
typing_extensions
